# PERF: groupby.any/all with object dtype