- [x] [Tests added and passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#writing-tests) if fixing a bug or adding a new feature
- [x] All [code checks passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#pre-commit).
- [x] Added [type annotations](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#type-hints) to new arguments/methods/functions.
- [x] Added an entry in the latest `doc/source/whatsnew/v2.0.0.rst` file if fixing a bug or adding a new feature.


```
from asv_bench.benchmarks.groupby import GroupByMethods

params = dict(dtype="object", method="all", application="direct", ncols=1) 

b = GroupByMethods()
b.setup(**params)

%timeit b.time_dtype_as_field(**params)

# 881 µs ± 20.2 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)  -> main
# 216 µs ± 1.15 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)  -> PR
```